requirements.txt
numpy
Pillow
rich
scikit-image
tqdm
ttkthemes
